02-physicsnemo-mgn: Phase-1 cross-review absorption — Finding 4 (NPZ loader fp32 preservation)
Finding 4 [Medium, cell 2]: load_mesh_rollout_npz previously coerced
node_positions + node_values to fp64 via `np.asarray(v, dtype=float)`,
silently contradicting SCHEMA.md §2's fp32 contract and Task 12's
`_assert_loader_contract_mgn` fp32 assertion. A well-formed
mesh_rollout.npz saved with fp32 values would fail the loader-contract
helper if Phase 2 wired it in after the canonical NPZ load — the
classic layered-fail-open where the schema, the writer, the reader,
and the validator each assume something different.

Fix: drop the `dtype=float` upcast — preserve on-disk dtype. A buggy
materializer that writes fp64 is now caught by
`_assert_loader_contract_mgn` rather than masked by the loader's silent
promotion.

Test: test_mesh_rollout_npz_round_trip_preserves_fp32_dtype — saves an
MGN-shaped fp32 fixture, reloads, asserts node_positions and
node_values["velocity"] dtype stay fp32. The pre-existing round-trip
test (test_save_load_mesh_rollout_npz_round_trip) still passes because
it uses np.testing.assert_allclose (value-based, not dtype-strict);
new test makes the dtype contract explicit.

Suite: 236 → 237 passed (one new test).

+++ external_validation/_rollout_anchors/_harness/mesh_rollout_adapter.py
@@ -270,11 +270,22 @@ def load_mesh_rollout_npz(path: Path | str) -> MeshRollout:
         missing = required - set(data.files)
         if missing:
             raise KeyError(f"mesh_rollout.npz {p} missing required fields: {sorted(missing)}")
-        node_positions = np.asarray(data["node_positions"], dtype=float)
+        # Preserve on-disk dtype per SCHEMA.md §2 fp32 contract. The
+        # prior `dtype=float` upcast silently promoted fp32 stored
+        # arrays to fp64, contradicting the SCHEMA contract and Task 12's
+        # `_assert_loader_contract_mgn` fp32 assertion — a layered-fail-
+        # open path surfaced by Phase-1 cross-review Finding 4: a
+        # well-formed mesh_rollout.npz saved with fp32 values would fail
+        # the loader-contract helper if Phase 2 wired it in after the
+        # canonical NPZ load. Preserving on-disk dtype keeps the round-
+        # trip honest and lets a buggy materializer that writes fp64 be
+        # caught by `_assert_loader_contract_mgn` rather than masked by
+        # this upcast.
+        node_positions = np.asarray(data["node_positions"])
         node_type = np.asarray(data["node_type"])
         nv_obj = data["node_values"]
         node_values_raw = nv_obj.item() if hasattr(nv_obj, "item") else dict(nv_obj)
-        node_values = {k: np.asarray(v, dtype=float) for k, v in node_values_raw.items()}
+        node_values = {k: np.asarray(v) for k, v in node_values_raw.items()}
         dt_arr = data["dt"]
         dt = float(dt_arr.item() if hasattr(dt_arr, "item") else dt_arr)
         meta_obj = data["metadata"]
